Moved more types and files to solo, fixed get session by code
