+++ functions/package.json
@@ -15,7 +15,7 @@
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "faker": "4.1.0",
-    "firebase-admin": "^9.2.0",
+    "firebase-admin": "9.2.0",
     "firebase-functions": "^3.6.1",
     "fp-ts": "2.8.4",
     "io-ts": "^2.2.11",
