+++ functions/src/api/session/api/getSessionByCode.ts
@@ -1,5 +1,8 @@
 import { RequestHandler } from "express";
 import * as admin from "firebase-admin";
+import { pipe } from "fp-ts/lib/pipeable";
+import { foldGameSession } from "solo-lib/lib/session";
+import { normalizeQuery } from "solo-lib/lib/utils/firebase";
 
 type ReqParams = {
   code: string;
@@ -8,17 +11,61 @@ type ReqParams = {
 export const getSessionByCode: RequestHandler<ReqParams> = async (req, res) => {
   const { code } = req.params;
 
-  const sessionByCode = await admin
-    .firestore()
-    .collection("session")
-    .where("code", "==", code)
-    .get();
+  try {
+    const sessionByCode = await admin
+      .firestore()
+      .collection("session")
+      .where("code", "==", code)
+      .get();
 
-  const [sessionDoc] = sessionByCode.docs.values();
+    const [sessionDoc] = sessionByCode.docs;
+    const sessionId = sessionDoc.id;
 
-  const session = sessionDoc.data();
-  const playersDoc = sessionDoc.get("players");
-  const players = playersDoc.data();
+    if (!sessionDoc.exists) {
+      res.status(404).send({ message: `Session with code ${code} not found.` });
+      return;
+    }
 
-  return res.send({ ...session, players });
+    const session = sessionDoc.data();
+    const sessionRef = admin.firestore().collection("session").doc(sessionId);
+
+    const playersDoc = await sessionRef.collection("players").get();
+    const players = normalizeQuery<SessionPlayer>(playersDoc);
+
+    function returnSimpleSession() {
+      res.send({ ...session, players });
+    }
+
+    async function returnGameSession() {
+      const players = normalizeQuery<SessionPlayer>(playersDoc);
+      const progressionDoc = await sessionRef.collection("progression").get();
+
+      const progression = normalizeQuery<Progression>(progressionDoc);
+      const cemeteryDoc = await sessionRef.collection("cemetery").get();
+
+      const cemetery = normalizeQuery<Cemetery>(cemeteryDoc);
+
+      return res.send({
+        ...session,
+        players,
+        progression,
+        cemetery,
+      });
+    }
+
+    console.log(session);
+
+    pipe(
+      session as SessionStore,
+      foldGameSession({
+        whenNoGameSession: returnSimpleSession,
+        whenLobbySession: returnSimpleSession,
+        whenLoadingSession: returnSimpleSession,
+        whenGameStarted: await returnGameSession,
+      })
+    );
+  } catch (e) {
+    console.error(e);
+    res.send(e);
+  }
 };
