+++ models/general.d.ts
@@ -1,1 +1,2 @@
 declare type ID = { id: string };
+declare type Normalized<T> = Record<string, T>;
