+++ solo/package.json
@@ -10,6 +10,7 @@
     "build": "yarn tsc"
   },
   "devDependencies": {
+    "firebase-admin": "9.2.0",
     "prettier": "2.1.2",
     "typescript": "3.9.7"
   },
