--- web/src/model/Session.ts
@@ -1,1 +1,0 @@
-export type Normalized<T> = Record<string, T>;
